requirements.txt
mythic
argparse
pyyaml
